requirements.txt
newsapi_python==0.2.7
streamlit==1.29.0
streamlit_antd_components==0.3.1
